Modernize the crate
The predicates.c was run through the c2rust transpiler to generate the
new code. The macros are still expanded and the generated code is
inherently unsafe, but this time there are no more calls to
uninitialized! Additionally an out of bounds bug was detected by Miri,
but luckily it does not affect the outcome of the algorithm. This bug
was resolved and a miri test is added to the CI.